requirements.txt
delta-spark==2.0.0
